MAJOR refactoring, TDD ftw.
Removed the Test object, the user should handle his own root nodes the way he
pleases.

Improved the delete method in the queryset, it's way faster now, specially when
removing many nodes.

Added a numchild property to every node, it slows down the queries that alter
the trees a little, but read operations are much faster.

Removed the Monty Python reference :( (since it was in the Test class, now
gone)